requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Signal Processing (DWT)
pywavelets>=1.4.1

# Scientific Computing
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress Bars
tqdm>=4.65.0

# HTTP Requests (for data collection)
requests>=2.31.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Optional: Jupyter for experimentation
# jupyter>=1.0.0
# ipython>=8.14.0

# Optional: Tensorboard for logging
# tensorboard>=2.13.0
